refactor(cli): move attach session logic to lib/
Relocate the view/drive/passthrough PTY-attach session logic to
lib/attach-{view,drive,passthrough}.ts and strip the now-dead standalone
registerXCommands. `local agent attach` consumes the attachX helpers; the
commands/ tree no longer holds attach implementations.
